backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.29
python-dotenv==1.0.1
google-generativeai==0.7.1
pydantic==2.7.1